requirements.txt
tabulate
discord.py===1.6.0